+++ relay-dynamic-config/src/global.rs
@@ -183,13 +183,13 @@ pub struct Options {
     )]
     pub http_span_allowed_hosts: Vec<String>,
 
-    /// Whether or not relay should drop attachments submitted with transactions.
+    /// Disables Relay from sending replay-events to Snuba.
     #[serde(
-        rename = "relay.drop-transaction-attachments",
+        rename = "replay.relay-snuba-publishing-disabled",
         deserialize_with = "default_on_error",
         skip_serializing_if = "is_default"
     )]
-    pub drop_transaction_attachments: bool,
+    pub replay_relay_snuba_publish_disabled: bool,
 
     /// All other unknown options.
     #[serde(flatten)]