+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+**Breaking Changes**:
+
+- Only check for local project configs in static mode. ([#5057](https://github.com/getsentry/relay/pull/5057))
+
 **Features**:
 
 - Add InstallableBuild and SizeAnalysis data categories. ([#5084](https://github.com/getsentry/relay/pull/5084))